Add a smart "typed null" placeholder
This 'Nil' object knows its generic type at runtime, and can generate
proxy objects which implement the interfaces of that type, with
customizable behavior per interface method via callbacks.

The term 'nil' was chosen, despite lots of possibility for confusion
with related paradigms such as the Scala language, because it is very
short while loosely meaning the same thing as 'null', 'None', etc.